SimpleDistMesh.SimpleDistMesh

SimpleDistMesh.Mesh

SimpleDistMesh.find_bars

SimpleDistMesh.grid_uniform_triangles

SimpleDistMesh.rejection_method

SimpleDistMesh.relax_mesh!

Mesh

Methods for creating 2-dimensional meshes using the DistMesh method are contained in the module,

SimpleDistMesh.SimpleDistMesh — Module

SimpleDistMesh

Mesh generation using signed distance functions

Meshes are stored as a set of node positions and triangles in column-major format in Mesh. A 2-dimensional mesh can be generated by defining a signed distance function $\mbox{d:\mathbb{R}^2 \rightarrow \mathbb{R}}$ and desired truss element length $\mbox{h:\mathbb{R}^2 \rightarrow \mathbb{R}}$, and passing to the Mesh constructor.

SimpleDistMesh.Mesh — Type

Mesh

Mesh containing nodes positions and triangulation

Properties

nodes::Matrix (2, n_nodes) 2-dimensional position of nodes

triangulation::Matrix{Int} (3, n_edges) node indices defining the triangles of the mesh

Utilities

The following functions perform various steps of the DistMesh mesh relaxation algorithm.

SimpleDistMesh.relax_mesh! — Function

relax_mesh!(nodes, bars; ...)

Move the nodes of the mesh according the forces between nodes, and return the maximum distance moved by an iterior node.

SimpleDistMesh.grid_uniform_triangles — Function

grid_uniform_triangles(bounds, h0)

Return the nodes for a grid of equilaterial triangles with side-length h0.

SimpleDistMesh.rejection_method — Function

rejection_method(d, h, nodes)

Return interior nodes which pass the rejection method

SimpleDistMesh.find_bars — Function

find_bars(d, geps, nodes)

Using a Delaunay triangulation return the bars connecting the nodes of the mesh and the indices of the mesh triangles

Output

(bars, triangulation)

bars::Matrix{Int} (2, n_edges) indices of nodes defining bar connections

triangulation::Matrix{Int} (3, n_triangles) indices of nodes defining the triangles of the mesh