SimpleDistMesh

SimpleDistMesh.Mesh

SimpleDistMesh.Mesh

SimpleDistMesh.find_bars

SimpleDistMesh.grid_uniform_triangles

SimpleDistMesh.rejection_method

SimpleDistMesh.relax_mesh!

Signed distance functions

SimpleDistMesh — Module

SimpleDistMesh

Mesh generation using signed distance functions

SimpleDistMesh.Mesh — Type

Mesh

Mesh with node positions and triangulation ts

The columns of the nodes matrix provide the positions of points in the mesh, while the triangles (or tetrahedra) of the mesh defined by the columns of ts.

Examples

A mesh in 2D comprised of the single triangle (0, 0), (1, 0), (1, 1)

julia> mesh = Mesh{Float64, 2}( [ 0.0 1.0 0.0 0.0 0.0 1.0 ], [1; 2; 3] )

SimpleDistMesh.Mesh — Method

Mesh(d, h, h0, bounds, fixed_nodes; ...)

Construct a mesh using the signed distance function d and desired edge length function h.

The initial edge length is given by h0 and the mesh is constructed within the bounding box given by bounds. Fixed nodes of the mesh can be provided as columns of fixed_nodes.

Arguments

d::Function signed distance function (== 0 on the boundary)

h::Function preferred edge length function

h0 starting edge length

bounds::Matrix defines bounding box for mesh with columns defining the minimum and maximum for each axis. For example, bounds = [x_min y_min; x_max y_max]

Keyword Arguments

Fscale internal pressure of relaxation method

fixed_nodes::Matrix defines fixed points of the mesh (if any)

Δt time-step of Euler method for point relaxation

geps tolerance in geometry calculations

dptol stopping criterion for relaxation method. Terminate if all node advections move less than dptol.

ttol for computational efficiency, triangulations are not computed at every step of the mesh relaxation. Recompute the mesh triangulation if a point has moved by at least ttol since last triangulation.

max_iters maximum number of iterations

Outputs

Returns a Mesh

SimpleDistMesh.find_bars — Method

find_bars(d, geps, nodes)

Return the bars which connect the nodes of the mesh using Delaunay triangulation and geometrical tolerance geps. The bars are defined by the columns the returned matrix.

SimpleDistMesh.grid_uniform_triangles — Method

grid_uniform_triangles(bounds, h0)

Return nodes of a uniform grid of equilaterial triangles with side-length h0. Nodes are returned as a vector of vectors.

SimpleDistMesh.rejection_method — Method

rejection_method(d, h, nodes)

Return subject of nodes inside the domain which pass the rejection method

SimpleDistMesh.relax_mesh! — Method

relax_mesh!(nodes, ts; ...)

Perform one relaxation iteration on the nodes and mesh triangulation `ts and return the maximum distance moved by an iterior node.