SimpleDistMesh.sdf.Circle

SimpleDistMesh.sdf.Polygon

SimpleDistMesh.sdf.Rect

SimpleDistMesh.sdf.Rotation

SimpleDistMesh.sdf.SignedDistFunc

Base.intersect

Base.setdiff

Base.union

SimpleDistMesh.sdf.distance

SimpleDistMesh.sdf.triangle_area

SimpleDistMesh.sdf.winding_number

Signed distance functions

SimpleDistMesh.sdf.Circle — Type

Circle(p0, r)

Signed distance function for a circle with center p0 and radius r

SimpleDistMesh.sdf.Polygon — Type

Polygon(nodes)

Signed distance function for a polygon with 2-dimensional nodes as columns of a matrix

Examples

Create signed distance function for a triangle

julia> tri = Polygon([0 1 0; 0 0 1])
(::Polygon) (generic function with 1 method)

SimpleDistMesh.sdf.Rect — Type

Rect(x1, x2, y1, y2)

Signed distance function for a rectangle [x1, x2] × [y1, y2]

SimpleDistMesh.sdf.Rotation — Type

Rotation(ϕ, p0)

Rotation about point p0 counter-clockwise by angle ϕ (in radians)

Examples

Rotate point about the origin by π/2

julia> Rotation(π/2)([1.0, 0.0])
2-element Vector{Float64}:
 6.123233995736766e-17
 1.0

SimpleDistMesh.sdf.SignedDistFunc — Type

SignedDistFunc

Abstract type for all signed distance functions

Base.intersect — Method

intersect(f, g) returns the intersection of the f and g regions for signed distance functions

Base.setdiff — Method

setdiff(f, g) returns the region of f with g removed for signed distance functions

Base.union — Method

union(f, g) returns the union of the regions for signed distance functions

SimpleDistMesh.sdf.distance — Method

distance(poly, x)

Return the minimal 2-norm distance from x to the edges of the polygon poly

SimpleDistMesh.sdf.triangle_area — Method

triangle_area(a, b, c)

Return the signed area of the 2D triangle abc

SimpleDistMesh.sdf.winding_number — Method

winding_number(poly, x)

Return the winding number of polygon poly about point x