SimpleDistMesh.Circle

SimpleDistMesh.ImplicitRegion

SimpleDistMesh.ImplicitRegion

SimpleDistMesh.Polygon

SimpleDistMesh.Rect

SimpleDistMesh.Rotation

SimpleDistMesh.SignedDistFunc

Base.in

Base.intersect

Base.setdiff

Base.union

SimpleDistMesh.distance

SimpleDistMesh.implicit_project

SimpleDistMesh.triangle_area

SimpleDistMesh.winding_number

Signed distance functions

A signed distance function, $d(x)$, can be any scalar function in 2-dimensions for which $d(x) = 0$ defines a meaningful region. For convenience, we have provided several simple types for creating common shapes.

SimpleDistMesh.Circle — Type

Circle(p0, r)

Signed distance function for a circle with center p0 and radius r

SimpleDistMesh.Rect — Type

Rect(x1, x2, y1, y2)

Signed distance function for a rectangle [x1, x2] × [y1, y2]

SimpleDistMesh.Polygon — Type

Polygon(nodes)

Signed distance function for a polygon with 2-dimensional nodes as columns of a matrix

Examples

Create signed distance function for a triangle

julia> triangle = Polygon([0 1 0; 0 0 1])
Polygon([0, 0], [1, 0], [0, 1])

SimpleDistMesh.ImplicitRegion — Type

ImplicitRegion(f)

Signed distance function for the implicit region defined by f(x) = 0

SimpleDistMesh.ImplicitRegion — Method

ImplicitRegion(f, c) returns a signed distance function for the implicit region defined by f(x) = c

Examples

c = Circle([0.0, 0.0], 1.0)

Circle([0.0, 0.0], 1.0)

Signed distance functions can also be composed with coordinate transforms to define a range of different geoemtries. For convenience, we also provide a type for defining rotations.

SimpleDistMesh.Rotation — Type

Rotation(ϕ, p0)

Rotation about point p0 counter-clockwise by angle ϕ (in radians)

Examples

Rotate point about the origin by π/2

julia> Rotation(π/2)([1.0, 0.0])
2-element Vector{Float64}:
 6.123233995736766e-17
 1.0

All of these are subtypes of the abstract type,

SimpleDistMesh.SignedDistFunc — Type

SignedDistFunc

Abstract type for all signed distance functions

Combining regions

The regions represented by signed distance can also be combined using common set operations of unions, intersections and set differences.

Base.union — Function

union(f, g) returns the union of the regions for signed distance functions

Base.setdiff — Function

setdiff(f, g) returns the region of f with g removed for signed distance functions

Base.intersect — Function

intersect(f, g) returns the intersection of the f and g regions for signed distance functions

Utilities

To create signed distance functions for polygonal geometries we define several useful functions.

SimpleDistMesh.implicit_project — Function

implicit_project(f, x0; max_iters, α)

Return the projection of x0 onto the level set f(x) = 0 such that the displacement from x0 is parallel to the gradient, (x - x0) ⋅ ∇f(x) = 0

SimpleDistMesh.distance — Function

distance(poly, x)

Return the minimal 2-norm distance from x to the edges of the polygon poly

SimpleDistMesh.winding_number — Function

winding_number(poly, x)

Return the winding number of polygon poly about point x

Base.in — Function

in(x::Vector, poly::Polygon)

Return true if point x is inside polygon poly

SimpleDistMesh.triangle_area — Function

triangle_area(a, b, c)

Return the signed area of the 2D triangle abc